pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trck"
version = "0.0.0"
authors = [{ name = "Mahdi Mir", email = "imahdimir@gmail.com" }]
description = "Automate Time Track Google Sheet"
readme = "README.md"
dependencies = [
    "pandas",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "gspread",
    "mtok",
]
